requirements.txt
tk
python-docx
requests
PyPDF2
pdf2image
pillow
pypdf
